refactor(domain): introduce @athena/domain package with ports and types
Create a dedicated application/domain tier (`packages/domain`) with zero
runtime dependencies, organised by bounded context (curriculum, speech).
Move the persisted data shapes and repository ports out of
`packages/api/src/trpc.ts` and `types.ts` into the domain package, and
relocate `verbalizeSymbols` into the speech context. Declare `LanguageCode`
once as the shared source of truth for the speakable languages.

`@athena/api` now depends on `@athena/domain` and re-exports its types and
ports, so the published API surface (including `@athena/api/speakable-text`)
is unchanged.

Addresses F3 (layer boundaries smeared inside packages/api) and lays the
groundwork for D4 (explicit bounded contexts).
