+++ packages/api/package.json
@@ -21,6 +21,7 @@
     "format": "prettier --write \"src/**/*.ts\" \"*.{js,ts,json}\""
   },
   "dependencies": {
+    "@athena/domain": "workspace:*",
     "@trpc/server": "^11.17.0",
     "zod": "^4.3.6"
   },
