+++ packages/api/src/types.ts
@@ -1,42 +1,14 @@
-export interface Lecture {
-  id: string;
-  title: string;
-  description: string;
-}
-
-export interface LectureListItem extends Lecture {
-  chapterCount: number;
-  questionCount: number;
-}
-
-export interface Chapter {
-  id: string;
-  lectureId: string;
-  association: string;
-  order: number;
-}
-
-export interface Question {
-  id: string;
-  chapterId: string;
-  question: string;
-  answer: string;
-  order: number;
-  isAnnotated: boolean;
-}
-
-export interface SpeechResult {
-  audioData: string; // Base64-encoded WAV
-  duration: number; // Duration in milliseconds
-}
-
-export type SpeechFormat = 'text' | 'ssml';
-
-export interface SpeechService {
-  synthesize(
-    text: string,
-    language: 'de' | 'en',
-    format?: SpeechFormat,
-  ): Promise<SpeechResult>;
-  isConfigured(): boolean;
-}
+/**
+ * Re-exports the domain data shapes so existing `./types.js` imports inside
+ * this package keep working. The canonical definitions now live in
+ * `@athena/domain`; this file is a thin compatibility barrel.
+ */
+export type {
+  Chapter,
+  Lecture,
+  LectureListItem,
+  Question,
+  SpeechFormat,
+  SpeechResult,
+  SpeechService,
+} from '@athena/domain';
