+++ packages/domain/eslint.config.js
@@ -1,0 +1,4 @@
+import { config } from '@athena/eslint-config';
+
+/** @type {import("eslint").Linter.Config[]} */
+export default [...config];
