+++ packages/domain/src/curriculum/index.ts
@@ -1,0 +1,11 @@
+export type {
+  Chapter,
+  Lecture,
+  LectureListItem,
+  Question,
+} from './types.js';
+export type {
+  ChapterRepository,
+  LectureRepository,
+  QuestionRepository,
+} from './ports.js';
