+++ packages/domain/src/speech/index.ts
@@ -1,0 +1,7 @@
+export type {
+  LanguageCode,
+  SpeechFormat,
+  SpeechResult,
+  SpeechService,
+} from './types.js';
+export { verbalizeSymbols } from './verbalize-symbols.js';
