+++ packages/domain/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@athena/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": "src"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist", "**/*.test.ts"]
+}
